+++ ft_strncmp.c
@@ -6,13 +6,15 @@
 /*   By: cyildiri <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/09/25 14:44:10 by cyildiri          #+#    #+#             */
-/*   Updated: 2016/09/25 14:47:01 by cyildiri         ###   ########.fr       */
+/*   Updated: 2016/09/26 12:24:28 by cyildiri         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include <string.h>
+
 int ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-    int index;
+    size_t	index;
 
     index = 0;
     while (index < n && s1[index] == s2[index])
