+++ ft_strncpy.c
@@ -6,16 +6,17 @@
 /*   By: cyildiri <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/09/25 11:40:26 by cyildiri          #+#    #+#             */
-/*   Updated: 2016/09/26 11:36:40 by cyildiri         ###   ########.fr       */
+/*   Updated: 2016/09/26 12:22:43 by cyildiri         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include<string.h>
+#include "ft_lib.h"
+#include <stdlib.h>
 
 char    *ft_strncpy(char *dst, const char *src, size_t len)
 {
-    int length;
-    int index;
+    size_t	length;
+    size_t	index;
 
     length = ft_strlen(src);
     dst = (char *)malloc(sizeof(char) * (length + 1));
